Move Recent Articles feed to its own collapsible column
